skenner_opt/requirements.txt
PyQt6>=6.5.0
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
pywin32>=306
comtypes>=1.2.0
